requirements.txt
qdrant-client
sentence-transformers
ollama
pypdf
python-docx
PyPDF2
Pillow